requirements.txt
biopython==1.66
cssmin==0.2.0
eventlet==0.18.4
Flask==0.10.1
Flask-Assets==0.11
Flask-BasicAuth==0.2.0
futures==3.0.5
greenlet==0.4.9
gunicorn==19.4.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
msgpack-python==0.4.7
numpy==1.10.4
pandas==0.17.1
python-dateutil==2.4.2
pytz==2015.7
requests==2.8.1
six==1.10.0
webassets==0.11.1
Werkzeug==0.11.2
wheel==0.24.0
